GR: ClearAtStart, SimpleNonHardcodedChoco, RelativePlacementSkel
Now the canvas is cleared at the begining to make the presence of
canvas visible on the screen.

Changed to a simple chocolate graphic object, which inturn has its
dimensions picked from a variable, which can be changed later to
adapt to the total screen area.

Added initial skeleton for picking up the x and y to start drawing
from a global reference, which will be updated as objects are
drawn to the screen, to represent the location of last object
drawn on screen. draw_biscuits references same.