Add multichoice helper
This helps generate a set of wrong answers for a multichoice set,
inturn the same is shown as a list of buttons.

The correct answer is placed randomly in one of the positions.

It is capable of limiting the generated values to be within a
specified max value, as well as as to how much precision will
be there in the values.