ShowMultiChoice: Callback from multichoice btn click handlers
The callback passed to show_multichoice, will be called, when ever
any of the buttons generated by multichoice will be clicked.

Inturn the argument passed to callback will also include the
result selected by the user (by their action of pressing one of
the buttons).

The add button handler inturn passes the vhint_add as the callback
to show_multichoice.

The arguments passed to show_multichoice has been changed to be a
single java object/map, rather than seperate arguments to the func.
This gives the flexibility to the developer to decide which subset
of args to pass to show_multichoice, in a flexible way.

Misc: Finding it difficult to drift in and out of half asleepness,
while also chipping at this.