User: Track and show time taken by user
The time taken by the user to solve any given type of question
is tracked wrt min, max and avg; and inturn the same is shown
to the user.

Added the root user.js logic to support this.

Also updated add and multichoice/getanswer logics to link into the
user helper module for the same.